Map body level symbols to the right project.
